// lib/utils/types/Path.ts
export type TPetal =
  | "heart"
  | "star"
  | "flower"
  | "heartFlipped"
  | "rose"
  | "sunflower"
  | "daisy"
  | "tulip"
  | "lily"
  | "orchid"
  | "lotus"
  | "cherryBlossom"
  | "bluebell"
  | "marigold"
  | "flowerFlipped"
  | "teardrop"
  | "spade"
  | "triangle"
  | "flame"
  | "diamond"
  | "rounded"
  | "ellipse";